+++ src/main/java/com/umc/pureum/domain/battle/dto/response/BattleMyProfilePhotoRes.java
@@ -1,4 +1,4 @@
-package com.umc.pureum.domain.battle.dto.repsonse;
+package com.umc.pureum.domain.battle.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
