+++ src/main/java/com/umc/pureum/domain/battle/dto/response/GetBattleSentenceInterface.java
@@ -1,4 +1,4 @@
-package com.umc.pureum.domain.battle.dto.repsonse;
+package com.umc.pureum.domain.battle.dto.response;
 
 public interface GetBattleSentenceInterface {
     Long getBattleSentenceId();
