+++ src/main/java/com/umc/pureum/domain/battle/repository/BattleResultRepository.java
@@ -1,6 +1,6 @@
 package com.umc.pureum.domain.battle.repository;
 
-import com.umc.pureum.domain.battle.dto.repsonse.GetBattleResultInterface;
+import com.umc.pureum.domain.battle.dto.response.GetBattleResultInterface;
 import com.umc.pureum.domain.battle.entity.BattleResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
